Move show entity edit dialog to the entity overview controllers
