+++ src/edu/wesimulated/firstapp/view/PersonOverviewController.java
@@ -1,13 +1,22 @@
 package edu.wesimulated.firstapp.view;
 
+import java.io.IOException;
+
+import javafx.collections.FXCollections;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Label;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
+import javafx.scene.layout.AnchorPane;
+import javafx.stage.Modality;
+import javafx.stage.Stage;
 import edu.wesimulated.firstapp.MainApp;
 import edu.wesimulated.firstapp.model.PersonData;
+import edu.wesimulated.firstapp.model.RoleData;
 
 public class PersonOverviewController {
 	@FXML
@@ -29,6 +38,29 @@ public class PersonOverviewController {
 	private MainApp mainApp;
 
 	public PersonOverviewController() {
+	}
+
+	public boolean showPersonEditDialog(PersonData person) {
+		try {
+			FXMLLoader loader = new FXMLLoader();
+			loader.setLocation(MainApp.class.getResource("view/PersonEditDialog.fxml"));
+			AnchorPane page = (AnchorPane) loader.load();
+			Stage dialogStage = new Stage();
+			dialogStage.setTitle("Edit Person");
+			dialogStage.initModality(Modality.WINDOW_MODAL);
+			dialogStage.initOwner(this.mainApp.getPrimaryStage());
+			Scene scene = new Scene(page);
+			dialogStage.setScene(scene);
+			PersonEditController controller = loader.getController();
+			controller.setDialogStage(dialogStage);
+			controller.setPerson(person);
+			controller.setMainApp(this.mainApp);
+			dialogStage.showAndWait();
+			return controller.isOkClicked();
+		} catch (IOException e) {
+			e.printStackTrace();
+			return false;
+		}
 	}
 
 	private void showPersonDetails(PersonData person) {
@@ -63,7 +95,7 @@ public class PersonOverviewController {
 	@FXML
 	private void handleNewPerson() {
 		PersonData tempPerson = new PersonData();
-		boolean okClicked = mainApp.showPersonEditDialog(tempPerson);
+		boolean okClicked = this.showPersonEditDialog(tempPerson);
 		if (okClicked) {
 			this.mainApp.getPersonData().add(tempPerson);
 		}
@@ -73,7 +105,7 @@ public class PersonOverviewController {
 	private void handleEditPerson() {
 		PersonData selectedPerson = personTable.getSelectionModel().getSelectedItem();
 		if (selectedPerson != null) {
-			boolean okClicked = mainApp.showPersonEditDialog(selectedPerson);
+			boolean okClicked = this.showPersonEditDialog(selectedPerson);
 			if (okClicked) {
 				showPersonDetails(selectedPerson);
 			}
